pyproject.toml
[project]
name = "telegram-jadicts-bot"
version = "0.3.0"
description = "Telegram 日本語辞書 BOT"
authors = [{ name = "mkpoli", email = "mkpoli@mkpo.li" }]
dependencies = [
    "loguru==0.5.3",
    "regex==2020.11.13",
    "python-telegram-bot==13.3",
    "pykakasi==2.0.6",
    "jaconv",
    "single-source>=0.3.0",
]
requires-python = ">=3.7"
license = { text = "MIT" }

[build-system]
requires = ["pdm-pep517>=1.0.0"]
build-backend = "pdm.pep517.api"


[tool]
[tool.pdm]
[tool.pdm.scripts]
start = "python bot.py"
